Made cursor return to the "home base"
The home base idea was originated because the main problem with v1.0 is
that the mouse moves too interestingly.

Computer execution speed is really fast, it would only require 0.05ms to
move the mouse and simulate a click, as my computer reports.

So moving the mouse back to home base doesn't take long. This process
can't be seen with human eyes.

Tested with excel and picture. Ready for release.